requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
APScheduler>=3.10.4
python-dotenv>=1.0.1
gunicorn>=23.0.0
psycopg[binary]>=3.2.1
